refactor(Table): refactor components folder
