+++ src/components/DataTable/__tests__/TableSearchBar.spec.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { fireEvent, render } from '@testing-library/react';
 import { DataTableContext } from '@/components/DataTable/context/DataTableContext';
-import { TableSearchBar } from '@/components/DataTable/TableSearchBar';
+import { TableSearchBar } from '@/components/DataTable/components/TableSearchBar';
 
 const handleSearch = vi.fn();
