+++ src/components/DataTable/components/EditableRow.tsx
@@ -1,56 +1,17 @@
-import React, { FC, ReactElement, useMemo, useState } from 'react';
-import { useDataTableContext } from '@/components/DataTable/context/DataTableContext';
-import { ColumnDefs } from './types';
+import React, { FC, useState } from 'react';
+import { Formik } from 'formik';
+import { TableBodyCell } from '@/components/DataTable/components/TableBodyCell';
 import { MdOutlineEdit } from 'react-icons/md';
 import { GiCancel, GiConfirmed } from 'react-icons/gi';
-import { Formik } from 'formik';
+import { RowValues } from '../types';
 
-export const TableBody = () => {
-  const { data, columnDefs } = useDataTableContext();
-
-  if (!data.length) return <h1>Empty table</h1>;
-  return (
-    <>
-      {data.map((row, i) => (
-        <TableRow key={i} data={row} columnDefs={columnDefs} />
-      ))}
-    </>
-  );
-};
-
-interface TableRowProps {
-  data: unknown;
-  columnDefs: ColumnDefs[];
-}
-
-const TableRow: FC<TableRowProps> = ({ data, columnDefs }) => {
-  const { onRowEdit } = useDataTableContext();
-
-  const rowValues = useMemo(
-    () => columnDefs.map((def) => ({ value: data[def.field], field: def.field, type: def.type })),
-    [data, columnDefs]
-  );
-
-  if (onRowEdit) {
-    return <EditableRow onSubmit={onRowEdit} initialValues={{ ...(data as object) }} rowValues={rowValues} />;
-  }
-
-  return (
-    <div className={'table-row'}>
-      {rowValues.map(({ field, value }, i) => (
-        <Cell key={i}>{value}</Cell>
-      ))}
-    </div>
-  );
-};
-
-interface EditableRowProps {
+interface Props {
   onSubmit: (values: unknown) => void;
   initialValues: any;
-  rowValues: Array<{ field: string; value: string | number; type: ColumnDefs['type'] }>;
+  rowValues: RowValues[];
 }
 
-const EditableRow: FC<EditableRowProps> = ({ onSubmit, rowValues, initialValues }) => {
+export const EditableRow: FC<Props> = ({ onSubmit, rowValues, initialValues }) => {
   const [isEditing, setIsEditing] = useState(false);
 
   return (
@@ -62,7 +23,7 @@ const EditableRow: FC<EditableRowProps> = ({ onSubmit, rowValues, initialValues 
       {(formik) => (
         <div className={'table-row'}>
           {rowValues.map(({ field, value, type }, i) => (
-            <Cell key={i}>
+            <TableBodyCell key={i}>
               {isEditing ? (
                 <input
                   type={type}
@@ -75,7 +36,7 @@ const EditableRow: FC<EditableRowProps> = ({ onSubmit, rowValues, initialValues 
               ) : (
                 <>{value}</>
               )}
-            </Cell>
+            </TableBodyCell>
           ))}
           <div className={'table-cell border-b py-3.5 px-2 cursor-pointer'}>
             <div className={'flex gap-3 items-center'}>
@@ -109,12 +70,4 @@ const EditableRow: FC<EditableRowProps> = ({ onSubmit, rowValues, initialValues 
       )}
     </Formik>
   );
-};
-
-interface CellProps {
-  children: ReactElement;
-}
-
-const Cell: FC<CellProps> = ({ children }) => {
-  return <div className={'table-cell border-b py-3.5 px-2 cursor-pointer'}>{children}</div>;
 };
