+++ src/components/DataTable/components/index.ts
@@ -1,0 +1,5 @@
+export * from './EditableRow';
+export * from './TableBodyCell';
+export * from './TableBody';
+export * from './TableHeader';
+export * from './TableSearchBar';
