+++ src/components/DataTable/types/column.ts
@@ -1,4 +1,7 @@
+type ColType = 'text' | 'number';
+
 export interface ColumnDefs {
   headerName: string;
   field: string;
+  type: ColType;
 }
