+++ src/components/DataTable/types/index.ts
@@ -1,1 +1,2 @@
 export * from './column';
+export * from './row';
