+++ src/components/DataTable/types/row.ts
@@ -1,0 +1,7 @@
+import { ColumnDefs } from './column';
+
+export interface RowValues {
+  value: unknown;
+  field: string;
+  type: ColumnDefs['type'];
+}
